# docs/06-evidence/users/auth_users_openapi.json
{
  "components": {
    "schemas": {
...
      "ConsentItem": {
        "properties": {
          "consentType": {
            "title": "Consenttype",
            "type": "string"
          },
          "policyVersion": {
            "title": "Policyversion",
            "type": "string"
          },
          "recordedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordedat"
          },
          "recordedPolicyVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordedpolicyversion"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "enum": [
              "granted",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "consentType",
          "policyVersion",
          "status",
          "recordedAt",
          "source"
        ],
        "title": "ConsentItem",
        "type": "object"
      },
      "ConsentResponse": {
        "properties": {
          "consent": {
            "$ref": "#/components/schemas/ConsentItem"
          }
        },
        "required": [
          "consent"
        ],
        "title": "ConsentResponse",
        "type": "object"
      },
      "ConsentStatusResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConsentItem"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PrivacyMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "ConsentStatusResponse",
        "type": "object"
      },
      "ConsentUpdate": {
        "additionalProperties": false,
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "policyVersion": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Policyversion",
            "type": "string"
          }
        },
        "required": [
          "policyVersion",
          "granted"
        ],
        "title": "ConsentUpdate",
        "type": "object"
      },
      "DeletionAdminResponse": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/DataRequestItem"
          },
          "userId": {
            "title": "Userid",
            "type": "integer"
          },
          "userUid": {
            "title": "Useruid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "userUid",
          "request"
        ],
        "title": "DeletionAdminResponse",
        "type": "object"
      },
      "DeletionAnonymizeResponse": {
        "properties": {
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl"
          },
          "request": {
            "$ref": "#/components/schemas/DataRequestItem"
          },
          "userId": {
            "title": "Userid",
            "type": "integer"
          },
          "userUid": {
            "title": "Useruid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "userUid",
          "request",
          "avatarUrl"
        ],
        "title": "DeletionAnonymizeResponse",
        "type": "object"
      },
      "DeletionRequestCreate": {
        "additionalProperties": false,
        "properties": {
          "currentPassword": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Currentpassword",
            "type": "string"
          },
          "reasonCode": {
            "enum": [
              "privacy",
              "unused",
              "other"
            ],
            "title": "Reasoncode",
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "reasonCode"
        ],
        "title": "DeletionRequestCreate",
        "type": "object"
      },
...
      "PrivacyMeta": {
        "properties": {
          "contractVersion": {
            "title": "Contractversion",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "contractVersion"
        ],
        "title": "PrivacyMeta",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Auth Users Contract Baseline",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/me/privacy/consents": {
      "get": {
        "operationId": "consent_status_api_v1_users_me_privacy_consents_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consent Status",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/me/privacy/consents/{consent_type}": {
      "put": {
        "operationId": "update_consent_api_v1_users_me_privacy_consents__consent_type__put",
        "parameters": [
          {
            "in": "path",
            "name": "consent_type",
            "required": true,
            "schema": {
              "enum": [
                "privacy_policy",
                "agent_memory"
              ],
              "title": "Consent Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Consent",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/me/privacy/deletion-requests": {
      "post": {
        "operationId": "request_deletion_api_v1_users_me_privacy_deletion_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletionRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Deletion",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/me/privacy/deletion-requests/current": {
      "get": {
        "operationId": "current_deletion_request_api_v1_users_me_privacy_deletion_requests_current_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current Deletion Request",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/me/privacy/deletion-requests/{request_uid}": {
      "delete": {
        "operationId": "cancel_deletion_api_v1_users_me_privacy_deletion_requests__request_uid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "request_uid",
            "required": true,
            "schema": {
              "title": "Request Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Deletion",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/me/privacy/export": {
      "post": {
        "operationId": "export_data_api_v1_users_me_privacy_export_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Data",
        "tags": [
          "privacy"
        ]
      }
    },
...
    "/api/v1/users/privacy/deletion-requests/{request_uid}/anonymize": {
      "post": {
        "operationId": "anonymize_deletion_api_v1_users_privacy_deletion_requests__request_uid__anonymize_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_uid",
            "required": true,
            "schema": {
              "title": "Request Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionAnonymizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anonymize Deletion",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/privacy/deletion-requests/{request_uid}/execute": {
      "post": {
        "operationId": "execute_deletion_api_v1_users_privacy_deletion_requests__request_uid__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_uid",
            "required": true,
            "schema": {
              "title": "Request Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionAdminResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Deletion",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/users/privacy/deletion-requests/{request_uid}/restore": {
      "post": {
        "operationId": "restore_deletion_api_v1_users_privacy_deletion_requests__request_uid__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_uid",
            "required": true,
            "schema": {
              "title": "Request Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionAdminResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Deletion",
        "tags": [
          "privacy"
        ]
      }
    }
  }
}